requirements.txt
pdfminer.six
requests
boto3
botocore
Flask
requests
json
faker

